implement mentionUpdate notifs
